pyproject.toml
[project]
name = "optimizer"
version = "0.1.0"
description = "L2 — Service d'Optimisation BESS SGE (Tewa Solar)"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    # Optimisation
    "cvxpy>=1.5",
    "highspy>=1.7",
    "numpy>=1.26",
    # DB
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    # Config
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "pyright>=1.1",
]

[project.scripts]
optimizer = "optimizer.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/optimizer"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/optimizer --cov-report=term-missing"
